requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
seaborn
certifi
pymongo
pymongo[srv]==3.12
torch
transformers==4.37.2
tqdm
ta
fastapi
xgboost
lightgbm
joblib
uvicorn 
scikit-learn